Testing

+++ ftpfiletos3.py
@@ -7,15 +7,14 @@ def main():
     import sys
 
     def signal_handler(signal, frame):
-        print("caught interrupt: " + str(signal) + " - restarting processing.")
-        t = threading.Thread(target=read_log_file).start()
+        if signal == 1:
+            print("caught interrupt: " + str(signal) + " - restarting processing.")
+            t = threading.Thread(target=read_log_file).start()
+        elif signal == 9:
+            print("caught kill signal... exiting...")
+            os.remove("/tmp/ftpfilestos3.pid")
+            sys.exit()
     # end signal_handler
-
-    def signal_kill_handler(signal, frame):
-        print("caught kill signal... exiting...")
-        os.remove("/tmp/ftpfilestos3.pid")
-        sys.exit()
-    # end signal_kill_handler
 
     pid = str(os.getpid())
     pidfile = "/tmp/ftpfilestos3.pid"
@@ -29,7 +28,7 @@ def main():
     # end with
 
     signal.signal(signal.SIGHUP, signal_handler)
-    signal.signal(signal.SIGKILL, signal_kill_handler)
+    signal.signal(signal.SIGKILL, signal_handler)
 
     t = threading.Thread(target=read_log_file).start()
